pyproject.toml
[project]
name = "linkedin_automation"
version = "0.1.0"
description = ""
authors = [
    {name = "Artem Danilov",email = "atlas.monker@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "selenium (==4.18.1)",
    "undetected-chromedriver (==3.5.5)",
    "python-dotenv (==1.0.1)",
    "fake-useragent (==1.5.1)",
    "requests (==2.31.0)",
    "coloredlogs (==15.0.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
